requirements.txt
streamlit>=1.20.0
openai>=1.0.0
python-dotenv>=0.21.0
bcrypt==4.2.0
pandas>=2.0.0